BayesNegativeBinomial

Documentation for BayesNegativeBinomial.

BayesNegativeBinomial.Sampler

BayesNegativeBinomial.sample

BayesNegativeBinomial.step!

BayesNegativeBinomial.Sampler — Type

Sampler(y::Vector{Int}, X::Matrix{Float64}; kwargs...)

Initialize a Gibbs Sampler for the following Bayesian Negative-Binomial model:

\[\begin{aligned} y_i | x_i, \beta, \gamma &\sim \text{NB}(r_{0y}, 1 / (1 + e^{- x_{i\gamma}'\beta_{\gamma}})), \\ \beta_k | \gamma_k &\sim \begin{cases} \mathcal{N}(\mu_{0\beta}, \Sigma_{0\beta}), &\text{if } \gamma = 1, \\ \delta_0 &\text{if } \gamma = 0, \end{cases} \\ \gamma &\sim \text{Womack}(\eta_{0\gamma}), \\ s &\sim \text{Gamma}(a_{0s}, b_{0s}), \end{aligned}\]

given a response vector y and a design matrix X. This constructor does not copy any of its arguments. Hence, for example, if y[1] is changed, it will affect the Gibbs sampler.

where $y_i \in \mathbb{N}_0$ is the response for the $i$-observation, $x_i \in \mathbb{R}^K$ is the associated covariate vector, $\delta_0(\cdot)$ is the Dirac measure centred at $0$ and $\text{Womack}(\eta_0)$ is the Womack distribution with parameter $\eta_{0\gamma}$ on $\{0, 1\}^K$.

Keyword arguments

β = zeros(size(X, 2)): current state of $\beta$.

Σ0β = 10 * I(size(X, 2)): $\Sigma_{0\beta}$.

μ0β = zeros(size(X, 2)): $\mu_{0\beta}$.

a0s = 1.0: $a_{0s}$.

b0s = 1.0: $a_{0s}$.

Example

julia> using Random, Distributions
julia> rng = MersenneTwister(1)
julia> X = randn(rng, 100, 2)
julia> y = rand(rng, 0:2, 10)
julia> s = BayesNegativeBinomial.Sampler(y, X)

BayesNegativeBinomial.sample — Method

sample(rng::AbstractRNG, s::BayesNegativeBinomial.Sampler; kwargs...)

Draw a posterior sample using the Gibbs sampler s, following Polson et al. (2013).

Arguments

mcmcsize = 2000: the total number of MCMC iterations.

burnin = 1000: the number of burn-in iterations.

Example

julia> using Random                 
julia> rng = MersenneTwister(1)     
julia> X = randn(rng, 100, 2)
julia> y = rand(rng, 0:2, 10)        
julia> sampler = BayesNegativeBinomial.Sampler(y, X)        
julia> chain = BayesNegativeBinomial.sample(rng, sampler)

References

Polson, N., Scott, J. & Windle, J. (2013) Bayesian inference for logistic models using Pólya–Gamma latent variables, Journal of the American Statistical Association, 108:504, 1339-1349, https://doi.org/10.1080/01621459.2013.829001.

BayesNegativeBinomial.step! — Method

step!(rng::AbstractRNG, s::BayesNegativeBinomial.Sampler)

Perform 1 iteration of the Gibbs sampler s, following Polson et al. (2013).

Example

julia> using Random, Distributions
julia> rng = MersenneTwister(1)
julia> X = randn(rng, 100, 2)
julia> y = rand(rng, 0:2, 10)
julia> s = BayesNegativeBinomial.Sampler(y, X)
julia> BayesNegativeBinomial.step!(rng, s)

References

Polson, N., Scott, J. & Windle, J. (2013) Bayesian inference for logistic models using Pólya–Gamma latent variables, Journal of the American Statistical Association, 108:504, 1339-1349, https://doi.org/10.1080/01621459.2013.829001.